Spatial Navigation: Add a scrollIntoView call when focusing an element. https://bugs.webkit.org/show_bug.cgi?id=36020
Reviewed by Simon Fraser.
Patch by Antonio Gomes <tonikitoo@webkit.org>

WebCore:

When focusing an element in Spatial Navigation logic, it is desired to make
this element visible in the current viewport. For that to happen, there
could be a call to Element's scrollIntoView method at focusing time. However
for visual aspects, it is preferable to scroll to an inflated rect of |element|
in order to consider the focus outline width.
As Element's scrollIntoView method does not provide this flexibility, patch adds
a custom scrollIntoView method to SpatialNavigation.h .

* page/FocusController.cpp:
(WebCore::FocusController::advanceFocusDirectionally):
* page/SpatialNavigation.cpp:
(WebCore::scrollIntoView):
* page/SpatialNavigation.h:

LayoutTests:

* platform/gtk/Skipped: Unskip snav-iframe-with-offscreen-element.html

Identifier: 47352@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56057 268f45cc-cd09-0410-ab3c-d52691b4dbfc