[Gtk+] Avoid making a copy of the HTTPHeaderMap.
ResourceRequest::httpHeaderFields() is returning a const
reference but we are storing it as a simple value. This
means that we run either the copy constructor or the assignment
operator and make a copy of the RefPtrs.

This extra copy can be avoided easily by storing
const reference.

* platform/network/soup/ResourceRequestSoup.cpp:
(WebCore::ResourceRequest::toSoupMessage):
* platform/network/soup/ResourceResponseSoup.cpp:
(WebCore::ResourceResponseSoup::toSoupMessage):

Identifier: 47442@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56148 268f45cc-cd09-0410-ab3c-d52691b4dbfc