WebCore: <rdar://problem/7761400> Rework the fix for https://bugs.webkit.org/show_bug.cgi?id=18722
Reviewed by Darin Adler.

Test: fast/dynamic/float-remove-above-line-2.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::removeFloatingObject): Treat 0- and less-than-0-height floats
as having a height of 1 so that they intersect with the line they originate on.
(WebCore::RenderBlock::clearFloats): Use numeric_limits.
* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::layoutInlineChildren): Removed the intersection checks here,
so that a float is always included in the float list of the line it originates on, even
if it does not intersect that line. This ensures that every float is accounted for, which
is necessary during incremental layout when adding floats from clean lines.

LayoutTests: Additional test for <rdar://problem/7761400>

Reviewed by Darin Adler.

* fast/dynamic/float-remove-above-line-2.html: Copied from LayoutTests/fast/dynamic/float-remove-above-line.html.
* fast/dynamic/float-remove-above-line-expected.txt: Updated.
* fast/dynamic/float-remove-above-line.html: Fixed inconsequential typo in end tag.



Identifier: 47480@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56186 268f45cc-cd09-0410-ab3c-d52691b4dbfc