Remove Geolocation.lastPosition, no longer in the spec. https://bugs.webkit.org/show_bug.cgi?id=36255 rdar://problem/7746357
Reviewed by Kenneth Rohde Christiansen.

* WebCore.base.exp: Updated since Geolocation's destructor is now non-virtual.

* page/Geolocation.cpp:
(WebCore::Geolocation::lastPosition): Add an assertion; it's only legal to
call this if access to the location is allowed.

* page/Geolocation.h: Removed unneeded includes. Made destructor non-virtual,
although it will still be virtual if any of the base classes have a virtual
destructor. Made lastPosition, isAllowed, and isDenied functions private.
Removed unused suspend, resume, setShouldClearCache, shouldClearCache,
and frame functions.

* page/Geolocation.idl: Removed lastPosition read-only attribute. No longer in
the Geolocation specification.



Identifier: 47482@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56188 268f45cc-cd09-0410-ab3c-d52691b4dbfc