<http://webkit.org/b/36560> resolve-ChangeLogs: git-rebase fails when resolve-ChangeLogs can't merge
Reviewed by Eric Seidel.

When resolve-ChangeLogs fails to merge a patch while running as
a git merge driver, it deletes the original file, which causes
an internal failure and stops git mid-merge:

    fatal: Failed to execute internal merge

The fix is to use the --force switch with patch so that it will
always attempt to apply the patch.  (The change in
mergeChangeLogs() for the previous commit also fixed this, but
adding --force also prevents any potential user interaction that
patch may want to display.)

* Scripts/VCSUtils.pm:
(mergeChangeLogs): Added --force switch to patch command.  Also
changed to use the exit status from the patch command to
determine the return value for this method.
* Scripts/webkitperl/VCSUtils_unittest/mergeChangeLogs.pl: Added
test to cover this bug.

Identifier: 47766@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56472 268f45cc-cd09-0410-ab3c-d52691b4dbfc