[Qt][GTK] http/tests/xmlhttprequest/xmlhttprequest-onProgress-open-should-zero-length.html fails after r56394 https://bugs.webkit.org/show_bug.cgi?id=36531
Reviewed by Alexey Proskuryakov.

The test was assuming that we would receive 3 progress events which may not be the case anymore
due to the event throttling added in r56394.

* http/tests/xmlhttprequest/xmlhttprequest-onProgress-open-should-zero-length.html: Removed
the counting as it belonged to the original version of the test and instead use the position
to determine when to reset the XHR.

* platform/gtk/Skipped:
* platform/qt/Skipped:
Removed the test from the Skipped lists now that it should pass.

Identifier: 47844@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@56550 268f45cc-cd09-0410-ab3c-d52691b4dbfc