<rdar://problem/7873647> Crash when updating hover state
Reviewed by Simon Fraser.

WebCore:

Test: fast/dynamic/hover-style-recalc-crash.html

Updating the hover state of an element caused the document to need style
recalc, and then updating the hover state of a link caused style recalc,
which changed the render tree while updateHoverActiveState() was iterating
over it, leading to a crash.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateHoverActiveState): Collect the nodes to be
updated into vectors, then update their active and hover states.

LayoutTests:

* fast/dynamic/hover-style-recalc-crash-expected.txt: Added.
* fast/dynamic/hover-style-recalc-crash.html: Added.



Identifier: 49040@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@57759 268f45cc-cd09-0410-ab3c-d52691b4dbfc