Crash while handling SVG font in the wrong namespace imported with @font-face https://bugs.webkit.org/show_bug.cgi?id=18862
Reviewed by Darin Adler.

WebCore:

Test: fast/invalid/invalidSVGFont.html

* loader/CachedFont.cpp:
(WebCore::CachedFont::getSVGFontById): Make sure we really get an SVGFontElement by using
getElementsByNameNS: the element factory chooses which element to create based on both
localName and namespace.

LayoutTests:

Test that load an SVG document not in the SVG namespace with @font-face does not
lead to a crash.

* fast/invalid/invalidSVGFont-expected.txt: Added.
* fast/invalid/invalidSVGFont.html: Added.
* fast/invalid/resources/svgFontWrongNamespace.svg: Added.

Identifier: 49060@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@57779 268f45cc-cd09-0410-ab3c-d52691b4dbfc