[Qt] Fix rendering of -webkit-user-select: none
Reviewed by Simon Hausmann.

-webkit-user-select: none is implemented by filling
the area with an invalid default-constructed Color.
In most ports passing an invalid color down to the
graphics backend seems to produce transparent fills.

In Qt the behavior of painting with an invalid QColor
is undefined, and in practice it results in painting
black opaque areas.

One way to fix this would be to use Qt::transparent
when converting an undefined Color to a QColor, but
Qt does not have short circuits for fully transparent
painting, and we actually end up in slow code paths
due to the transparency. So, we're better of doing the
short circuit in WebKit.

https://bugs.webkit.org/show_bug.cgi?id=38523

* platform/graphics/qt/GraphicsContextQt.cpp:

Identifier: 50195@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@58949 268f45cc-cd09-0410-ab3c-d52691b4dbfc