Allows build-webkit script to receive an install prefix as parameter https://bugs.webkit.org/show_bug.cgi?id=26224
Reviewed by Kenneth Christiansen , Simon Hausmann and Gustavo Noronha.
Patch by Rodrigo Belem <rodrigo.belem@openbossa.org>

WebCore:

This patch adds the ability, in the QtWebkit build system, to change
the installation path.

* WebCore.pro:

WebKitTools:

Added more parameters to build-webkit script, the --prefix for gkt
and --install-libs, --install-headers for qt. Now it is possible
to change the install prefix for gtk and install path for qt.

* Scripts/build-webkit:
* Scripts/webkitdirs.pm:

Identifier: 50473@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@59235 268f45cc-cd09-0410-ab3c-d52691b4dbfc