Adding block elements to a block and removing them changes the height of the element. https://bugs.webkit.org/show_bug.cgi?id=38874
Reviewed by Dave Hyatt.

WebCore:

The initial height of an empty editable block is different from the height of the same block after we add block elements
and remove them.

Test: editing/deleting/delete-block-contents-004.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::removeChild): m_childrenInline was set to true only when the Renderer is created
and never reset to true when all the block elements contained are removed.

LayoutTests:

* editing/deleting/delete-block-contents-004-expected.txt: Added.
* editing/deleting/delete-block-contents-004.html: Added.
The following tests have new results because of the height change and because we now have one less renderer after the DOM changes.
* editing/deleting/in-visibly-empty-root-expected.txt:
* platform/mac/editing/deleting/5408255-expected.txt:
* platform/mac/editing/deleting/delete-select-all-001-expected.txt:
* platform/mac/editing/selection/4895428-1-expected.txt:
* platform/mac/editing/selection/4895428-4-expected.txt:



Identifier: 50483@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@59245 268f45cc-cd09-0410-ab3c-d52691b4dbfc