Fix an HRGN leak in WKCACFLayerRenderer
Fixes <http://webkit.org/b/39312> <rdar://problem/7998728> REGRESSION
(r53686-r55990): Graphics corruption when watching video (affects
youtube.com)

Reviewed by Darin Adler.

* platform/graphics/win/WKCACFLayerRenderer.cpp:
(WebCore::getDirtyRects): Use an OwnPtr to hold the HRGN we allocate.
That way we don't have to remember to destroy it (as we were
forgetting to do) when we bail out of this function early.

Identifier: 50973@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@59864 268f45cc-cd09-0410-ab3c-d52691b4dbfc