Fixed failure seen on buildbot (due to flaky test).
Reviewed by Oliver Hunt.

Don't report an exception in a nested script evaluation -- just pass the
exception on to our caller. Our caller can decide to report the exception,
if it fully propagates -- but we don't want to report an exception that's
going to get caught.

* bindings/js/ScriptController.cpp:
(WebCore::ScriptController::ScriptController):
(WebCore::ScriptController::evaluateInWorld):
* bindings/js/ScriptController.h:



Identifier: 51036@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@59943 268f45cc-cd09-0410-ab3c-d52691b4dbfc