REGRESSION: (r47291): Upload progress events are not fired for simple cross-site XHR. https://bugs.webkit.org/show_bug.cgi?id=39029
Reviewed by Alexey Proskuryakov.

WebCore:

Specifically, WebKit should fire upload events if one or more event listeners are
registered on the XMLHttpRequestUpload object when send is called in an async manner.

* xml/XMLHttpRequest.cpp:
(WebCore::XMLHttpRequest::createRequest): Allow upload events to be fired when there are
handlers for them in the cross-origin simple request case.

LayoutTests:

* http/tests/xmlhttprequest/simple-cross-origin-progress-events-expected.txt: Update results.
* http/tests/xmlhttprequest/simple-cross-origin-progress-events.html: Add a test which adds
  an upload progress event handler before XMLHttpRequest.send() is called.

Identifier: 52661@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@61654 268f45cc-cd09-0410-ab3c-d52691b4dbfc