Speculative fix for <rdar://problem/8071558> CrashTracer: [USER] 2300+ crashes in Safari at com.apple.WebCore: WebCore::FrameView::scheduleRelayout + 352
Reviewed by Darin Adler.

Unfortunately, we don't have a reproducible case for this bug, and
therefore, we do not have a layout test either. It is pretty clear
from the logs that m_frame->settings() is null in
FrameView::scheduleRelayout() in the crashing case.
m_frame->settings() is null whenever page is null. Everywhere else
in FrameView.cpp we null-check either page or settings before using
settings. It seems plausible to me that scheduleRelayout could be
called when page is null, so the fix is just to add null-checks.

* page/FrameView.cpp:
(WebCore::FrameView::layout):
(WebCore::FrameView::scheduleRelayout):



Identifier: 53093@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@62135 268f45cc-cd09-0410-ab3c-d52691b4dbfc