optimizeLegibility doesn't play nice with fonts that do not have a space glyph https://bugs.webkit.org/show_bug.cgi?id=41599
Reviewed by Sam Weinig.

No test because none of the fonts available to DumpRenderTree are missing a space glyph.

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::findNextLineBreak): When the font does not map the space character
to a glyph, a fallback font is used for space. Therefore, wordTrailingSpaceWidth must be
initialized with the width of a space as measured by the Font rather than with the cached
space width.



Identifier: 53425@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@62510 268f45cc-cd09-0410-ab3c-d52691b4dbfc