Fix clean builds on Mac OS X.
Since "Copy Forwarding and ICU Headers" was split in to a separate target
from the WebCore framework it no longer had automatica access to the
PRIVATE_HEADERS_FOLDER_PATH variable. This caused it to copy the header
files in to the wrong location. Fix that by giving PRIVATE_HEADERS_FOLDER_PATH
its correct value and ensuring that the path exists.

* WebCore.xcodeproj/project.pbxproj:

Identifier: 53531@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@62616 268f45cc-cd09-0410-ab3c-d52691b4dbfc