Use ::MessageBox instead of ::DebugBreak to make the web process wait for a debugger
While ::DebugBreak can make it easier to attach a debugger (because it
brings up the Just-in-Time debugger window), using ::MessageBox has
two advantages:
  1) It explains to the user what's going on, so it's harder to
     mistake this debugging aid for a bug.
  2) It allows you to debug the UI process and the web process using
     the same debugger, if desired.

Special thanks to John Sullivan for help with the text that's shown in
the alert.

Fixes <http://webkit.org/b/42848> WebKit2's wait-for-debugger
debugging aid should explain to the user what's happening

Reviewed by John Sullivan and Anders Carlsson.

* WebProcess/WebKitMain.cpp:
(WebKitMain): Use a ::MessageBox instead of a ::DebugBreak to wait for
the debugger. The alert explains to the user what's going on.

Identifier: 54750@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@63913 268f45cc-cd09-0410-ab3c-d52691b4dbfc