It should be possible to attach to the WebProcess in gdb during startup https://bugs.webkit.org/show_bug.cgi?id=42853
Reviewed by Adam Roben.

Create an unique bootstrap service name and pass it to the web process, instead of registering a per process
name (which was incorrect anyway). This lets us get rid of the requirement that the UI process always should be
the parent process of the web process, something which is false when running under the debugger.

* UIProcess/Launcher/mac/ProcessLauncherMac.mm:
(WebKit::ProcessLauncher::launchProcess):
* WebProcess/mac/WebProcessMainMac.mm:
(WebKit::WebProcessMain):

Identifier: 54754@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@63917 268f45cc-cd09-0410-ab3c-d52691b4dbfc