[Windows] Home hey doesn't work in first DIV inside a TABLE https://bugs.webkit.org/show_bug.cgi?id=42767
Reviewed by Ojan Vafai.

WebCore:

The bug was caused by positionAvoidingFirstPositionInTable moving outside of the editing region.
Added a check so that it avoids the first position in the table only if the table is editable.

Test: editing/selection/home-inside-noneditable-table.html

* editing/visible_units.cpp:
(WebCore::positionAvoidingFirstPositionInTable):

LayoutTests:

Added a test to move caret backwards to the line boundary inside a content-editable div
inside a non-editable table.

* editing/selection/home-inside-noneditable-table-expected.txt: Added.
* editing/selection/home-inside-noneditable-table.html: Added.
* resources/dump-as-markup.js: Modified to support multiple calls to dump
(Markup.dump.get if):
(Markup.dump):
(Markup.notifyDone):



Identifier: 54755@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@63918 268f45cc-cd09-0410-ab3c-d52691b4dbfc